.history/requirements_20250416154611.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
matplotlib